Move Search apps into their own sub modules
