+++ datahub/search/apps.py
@@ -1,7 +1,7 @@
 from django.apps import AppConfig
 from django.conf import settings
 
-from datahub.search import elasticsearch, models
+from datahub.search import elasticsearch
 
 
 class SearchConfig(AppConfig):
@@ -12,15 +12,21 @@ class SearchConfig(AppConfig):
 
     def ready(self):
         """Configures Elasticsearch default connection."""
+        from .company.models import Company
+        from .contact.models import Contact
+        from .investment.models import InvestmentProject
+
         elasticsearch.configure_connection()
         elasticsearch.configure_index(settings.ES_INDEX)
 
         # Makes sure mappings exist in Elasticsearch.
         # Those calls are idempotent
-        models.Company.init(index=settings.ES_INDEX)
-        models.Contact.init(index=settings.ES_INDEX)
-        models.InvestmentProject.init(index=settings.ES_INDEX)
+        Company.init(index=settings.ES_INDEX)
+        Contact.init(index=settings.ES_INDEX)
+        InvestmentProject.init(index=settings.ES_INDEX)
 
-        # Let's import Company and Contact post_save signal handlers
+        # Let's import all post_save signal handlers
         # So DB models can be synced with Elasticsearch on save
-        import datahub.search.signals  # noqa
+        import datahub.search.company.signals  # noqa
+        import datahub.search.contact.signals  # noqa
+        import datahub.search.investment.signals  # noqa
