+++ datahub/search/dict_utils.py
@@ -1,0 +1,38 @@
+def _id_name_dict(obj):
+    """Creates dictionary with selected field from supplied object."""
+    return {
+        'id': str(obj.id),
+        'name': obj.name,
+    }
+
+
+def _id_type_dict(obj):
+    """Creates dictionary with selected field from supplied object."""
+    return {
+        'id': str(obj.id),
+        'type': obj.type
+    }
+
+
+def _id_uri_dict(obj):
+    """Creates dictionary with selected field from supplied object."""
+    return {
+        'id': str(obj.id),
+        'uri': obj.uri
+    }
+
+
+def _contact_dict(obj):
+    """Creates dictionary with selected field from supplied object."""
+    return {
+        'id': str(obj.id),
+        'first_name': obj.first_name,
+        'last_name': obj.last_name,
+    }
+
+
+def _company_dict(obj):
+    return {
+        'id': str(obj.id),
+        'company_number': obj.company_number,
+    }
